// services/admin-ui/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        RecordCreate: {
            /**
             * Type
             * @enum {string}
             */
            type: "individual" | "bulk";
            /** Branch Id */
            branch_id: string;
            /** Name */
            name: string;
            /** Org Id */
            org_id?: string | null;
            /** Participant Id */
            participant_id?: number | null;
            /**
             * Minutes
             * @description Minutes of meditation practice
             */
            minutes: number;
            /** Participant Count */
            participant_count?: number | null;
            /** Minutes Per Person */
            minutes_per_person?: number | null;
            /**
             * Morning Male
             * @default 0
             */
            morning_male: number;
            /**
             * Morning Female
             * @default 0
             */
            morning_female: number;
            /**
             * Morning Unspecified
             * @default 0
             */
            morning_unspecified: number;
            /**
             * Afternoon Male
             * @default 0
             */
            afternoon_male: number;
            /**
             * Afternoon Female
             * @default 0
             */
            afternoon_female: number;
            /**
             * Afternoon Unspecified
             * @default 0
             */
            afternoon_unspecified: number;
            /**
             * Evening Male
             * @default 0
             */
            evening_male: number;
            /**
             * Evening Female
             * @default 0
             */
            evening_female: number;
            /**
             * Evening Unspecified
             * @default 0
             */
            evening_unspecified: number;
            /**
             * Date
             * Format: date
             */
            date: string;
            /** Photo Url */
            photo_url?: string | null;
            /** Latitude */
            latitude?: number | null;
            /** Longitude */
            longitude?: number | null;
            /** Submitted By */
            submitted_by?: string | null;
            /** Submitted Phone */
            submitted_phone?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}